Bump sqed to 0.8.0.  Stage/left-t backend. #3400. #3367
Better Sqed error handling?!

+++ app/models/image.rb
@@ -81,14 +81,13 @@ class Image < ApplicationRecord
   # also using https://github.com/teeparham/paperclip-meta
   has_attached_file :image_file,
     styles: {
-    thumb: [ "#{DEFAULT_SIZES[:thumb][:width]}x#{DEFAULT_SIZES[:thumb][:height]}>", :png ] ,
-    medium: [ "#{DEFAULT_SIZES[:medium][:width]}x#{DEFAULT_SIZES[:medium][:height]}>", :jpg ]
-  },
-  default_url: MISSING_IMAGE_PATH,
-  filename_cleaner: Utilities::CleanseFilename,
-  processors: [:rotator]
+      thumb: [ "#{DEFAULT_SIZES[:thumb][:width]}x#{DEFAULT_SIZES[:thumb][:height]}>", :png ] ,
+      medium: [ "#{DEFAULT_SIZES[:medium][:width]}x#{DEFAULT_SIZES[:medium][:height]}>", :jpg ] },
+    default_url: MISSING_IMAGE_PATH,
+    filename_cleaner: Utilities::CleanseFilename,
+    processors: [:rotator]
 
-  #:restricted_characters => /[^A-Za-z0-9\.]/,
+    #:restricted_characters => /[^A-Za-z0-9\.]/,
   validates_attachment_content_type :image_file, content_type: /\Aimage\/.*\Z/
   validates_attachment_presence :image_file
   validate :image_dimensions_too_short
@@ -97,19 +96,19 @@ class Image < ApplicationRecord
 
   accepts_nested_attributes_for :sled_image, allow_destroy: true
 
-  # @return [Boolean]
+    # @return [Boolean]
   def has_duplicate?
     Image.where(image_file_fingerprint: self.image_file_fingerprint).count > 1
   end
 
-  # @return [Array]
+    # @return [Array]
   def duplicate_images
     Image.where(image_file_fingerprint: self.image_file_fingerprint).not_self(self).to_a
   end
 
-  # @return [Hash]
-  # returns a hash of EXIF data if present, empty hash if not.do
-  # EXIF data tags/specifications -  http://web.archive.org/web/20131018091152/http://exif.org/Exif2-2.PDF
+    # @return [Hash]
+    # returns a hash of EXIF data if present, empty hash if not.do
+    # EXIF data tags/specifications -  http://web.archive.org/web/20131018091152/http://exif.org/Exif2-2.PDF
   def exif
     ret_val = {} # return value
 
@@ -128,6 +127,7 @@ class Image < ApplicationRecord
     ret_val # return
   end
 
+  # TODO: move to /lib
   # @return [Nil]
   #  currently handling this client side
   def gps_data
@@ -151,17 +151,17 @@ class Image < ApplicationRecord
     # - maybe 2 versions? - one returns string, other decimal?
   end
 
-  # Returns the true, unscaled height/width ratio
-  # @return [Float]
+    # Returns the true, unscaled height/width ratio
+    # @return [Float]
   def hw_ratio
     raise if height.nil? || width.nil? # if they are something has gone badly wrong
     return (height.to_f / width.to_f)
   end
 
-  # rubocop:disable Style/StringHashKeys
-  # used in ImageHelper#image_thumb_tag
-  # asthetic scaling of very narrow images in thumbnails
-  # @return [Hash]
+    # rubocop:disable Style/StringHashKeys
+    # used in ImageHelper#image_thumb_tag
+    # asthetic scaling of very narrow images in thumbnails
+    # @return [Hash]
   def thumb_scaler
     a = self.hw_ratio
     if a < 0.6
@@ -170,23 +170,23 @@ class Image < ApplicationRecord
       {}
     end
   end
-  # rubocop:enable Style/StringHashKeys
+    # rubocop:enable Style/StringHashKeys
 
-  # the scale factor is typically the same except in a few cases where we skew small thumbs
-  # @param [Symbol] size
-  # @return [Float]
+    # the scale factor is typically the same except in a few cases where we skew small thumbs
+    # @param [Symbol] size
+    # @return [Float]
   def width_scale_for_size(size = :medium)
     (width_for_size(size).to_f / width.to_f)
   end
 
-  # @param [Symbol] size
-  # @return [Float]
+    # @param [Symbol] size
+    # @return [Float]
   def height_scale_for_size(size = :medium)
     height_for_size(size).to_f / height.to_f
   end
 
-  # @param [Symbol] size
-  # @return [Float]
+    # @param [Symbol] size
+    # @return [Float]
   def width_for_size(size = :medium)
     a = self.hw_ratio
     case size
@@ -203,8 +203,8 @@ class Image < ApplicationRecord
     end
   end
 
-  # @param [Symbol] size
-  # @return [Float]
+    # @param [Symbol] size
+    # @return [Float]
   def height_for_size(size = :medium)
     a = self.hw_ratio
     case size
@@ -221,23 +221,23 @@ class Image < ApplicationRecord
     end
   end
 
-  #  def filename(layout_section_type)
-  #    'tmp/' + tempfile(layout_section_type).path.split('/').last
-  #  end
+    #  def filename(layout_section_type)
+    #    'tmp/' + tempfile(layout_section_type).path.split('/').last
+    #  end
 
-  #  def tempfile(layout_section_type)
-  #    tempfile = Tempfile.new([layout_section_type.to_s, '.jpg'], "#{Rails.root.to_s}/public/images/tmp", encoding: 'ASCII-8BIT' )
-  #    tempfile.write(zoomed_image(layout_section_type).to_blob)
-  #    tempfile
-  #  end
+    #  def tempfile(layout_section_type)
+    #    tempfile = Tempfile.new([layout_section_type.to_s, '.jpg'], "#{Rails.root.to_s}/public/images/tmp", encoding: 'ASCII-8BIT' )
+    #    tempfile.write(zoomed_image(layout_section_type).to_blob)
+    #    tempfile
+    #  end
 
-  # @param [ActionController::Parameters] params
-  # @return [Magick::Image]
+    # @param [ActionController::Parameters] params
+    # @return [Magick::Image]
   def self.cropped(params)
     image = Image.find(params[:id])
     img = Magick::Image.read(image.image_file.path(:original)).first
     begin
-    # img.crop(x, y, width, height, true)
+      # img.crop(x, y, width, height, true)
       cropped = img.crop( params[:x].to_i, params[:y].to_i, params[:width].to_i, params[:height].to_i, true)
     rescue RuntimeError
       cropped = img.crop(0,0, 1, 1)  # return a single pixel on error ! TODO: make/return an error image
@@ -247,8 +247,8 @@ class Image < ApplicationRecord
     cropped
   end
 
-  # @param [ActionController::Parameters] params
-  # @return [Magick::Image]
+    # @param [ActionController::Parameters] params
+    # @return [Magick::Image]
   def self.resized(params)
     c = cropped(params)
     resized = c.resize(params[:new_width].to_i, params[:new_height].to_i) #.sharpen(0x1)
@@ -256,72 +256,75 @@ class Image < ApplicationRecord
     resized
   end
 
-  # @param [ActionController::Parameters] params
-  # @return [Magick::Image]
+    # @param [ActionController::Parameters] params
+    # @return [Magick::Image, nil]
   def self.scaled_to_box(params)
-    c = cropped(params)
-    ratio = c.columns.to_f / c.rows.to_f
-    box_ratio = params[:box_width].to_f / params[:box_height].to_f
-    # TODO: special considerations for 1:1?
+    begin
+      c = cropped(params)
+      ratio = c.columns.to_f / c.rows.to_f
+      box_ratio = params[:box_width].to_f / params[:box_height].to_f
+      # TODO: special considerations for 1:1?
 
-    if box_ratio > 1
-      if ratio > 1 # wide into wide
-        scaled = c.resize(
-          params[:box_width].to_i,
-          (params[:box_height].to_f / ratio * box_ratio).to_i
-        ) #.sharpen(0x1)
-      else # tall into wide
-        scaled = c.resize(
-          (params[:box_width ].to_f * ratio / box_ratio).to_i,
-          params[:box_height].to_i
-        ) #.sharpen(0x1)
+      if box_ratio > 1
+        if ratio > 1 # wide into wide
+          scaled = c.resize(
+            params[:box_width].to_i,
+            (params[:box_height].to_f / ratio * box_ratio).to_i
+          ) #.sharpen(0x1)
+        else # tall into wide
+          scaled = c.resize(
+            (params[:box_width ].to_f * ratio / box_ratio).to_i,
+            params[:box_height].to_i
+          ) #.sharpen(0x1)
+        end
+      else # <
+        if ratio > 1 # wide into tall
+          scaled = c.resize(
+            params[:box_width].to_i,
+            (params[:box_height].to_f / ratio * box_ratio).to_i
+          ) #.sharpen(0x1)
+        else # tall into tall # TODO: or 1:1?!
+          scaled = c.resize(
+            (params[:box_width ].to_f * ratio / box_ratio ).to_i,
+            (params[:box_height].to_f ).to_i
+          ) #.sharpen(0x1)
+        end
       end
-    else # <
-      if ratio > 1 # wide into tall
-        scaled = c.resize(
-          params[:box_width].to_i,
-          (params[:box_height].to_f / ratio * box_ratio).to_i
-        ) #.sharpen(0x1)
-      else # tall into tall # TODO: or 1:1?!
-        scaled = c.resize(
-          (params[:box_width ].to_f * ratio / box_ratio ).to_i,
-          (params[:box_height].to_f ).to_i
-        ) #.sharpen(0x1)
-      end
+      c.destroy!
+    rescue Magick::ImageMagickError
+      nil
     end
-    c.destroy!
-
     scaled
   end
 
-  # @param [ActionController::Parameters] params
-  # @return [String]
+    # @param [ActionController::Parameters] params
+    # @return [String]
   def self.scaled_to_box_blob(params)
     self.to_blob!(scaled_to_box(params))
   end
 
-  # @param [ActionController::Parameters] params
-  # @return [String]
+    # @param [ActionController::Parameters] params
+    # @return [String]
   def self.resized_blob(params)
     self.to_blob!(resized(params))
   end
 
-  # @param [ActionController::Parameters] params
-  # @return [String]
+    # @param [ActionController::Parameters] params
+    # @return [String]
   def self.cropped_blob(params)
     self.to_blob!(cropped(params))
   end
 
-  # @param used_on [String] required, a depictable base class name like  `Otu`, `Content`, or `CollectionObject`
-  # @return [Scope]
-  #   the max 10 most recently used images, as `used_on`
+    # @param used_on [String] required, a depictable base class name like  `Otu`, `Content`, or `CollectionObject`
+    # @return [Scope]
+    #   the max 10 most recently used images, as `used_on`
   def self.used_recently(user_id, project_id, used_on = '')
     i = arel_table
     d = Depiction.arel_table
 
     # i is a select manager
     j = d.project(d['image_id'], d['updated_at'], d['depiction_object_type']).from(d)
-      .where(d['updated_at'].gt( 1.weeks.ago ))
+      .where(d['updated_at'].gt( 1.week.ago ))
       .where(d['updated_by_id'].eq(user_id))
       .where(d['project_id'].eq(project_id))
       .order(d['updated_at'].desc)
@@ -335,39 +338,39 @@ class Image < ApplicationRecord
     joins(k).distinct.pluck(:id)
   end
 
-  # @params target [String] required, one of nil, `AssertedDistribution`, `Content`, `BiologicalAssociation`, 'TaxonDetermination'
-  # @return [Hash] images optimized for user selection
+    # @params target [String] required, one of nil, `AssertedDistribution`, `Content`, `BiologicalAssociation`, 'TaxonDetermination'
+    # @return [Hash] images optimized for user selection
   def self.select_optimized(user_id, project_id, target = nil)
     r = used_recently(user_id, project_id, target)
     h = {
       quick: [],
-      pinboard: Image.pinned_by(user_id).where(project_id: project_id).to_a,
+      pinboard: Image.pinned_by(user_id).where(project_id:).to_a,
       recent: []
     }
 
     if target && !r.empty?
       h[:recent] = (
         Image.where('"images"."id" IN (?)', r.first(5) ).to_a +
-        Image.where(project_id: project_id, created_by_id: user_id, created_at: 3.hours.ago..Time.now)
+        Image.where(project_id:, created_by_id: user_id, created_at: 3.hours.ago..Time.now)
         .order('updated_at DESC')
         .limit(3).to_a
       ).uniq.sort{|a,b| a.updated_at <=> b.updated_at}
 
       h[:quick] = (
-        Image.pinned_by(user_id).pinboard_inserted.where(project_id: project_id).to_a +
+        Image.pinned_by(user_id).pinboard_inserted.where(project_id:).to_a +
         Image.where('"images"."id" IN (?)', r.first(4) ).to_a)
         .uniq.sort{|a,b| a.updated_at <=> b.updated_at}
     else
-      h[:recent] = Image.where(project_id: project_id).order('updated_at DESC').limit(10).to_a
-      h[:quick] = Image.pinned_by(user_id).pinboard_inserted.where(pinboard_items: {project_id: project_id}).order('updated_at DESC')
+      h[:recent] = Image.where(project_id:).order('updated_at DESC').limit(10).to_a
+      h[:quick] = Image.pinned_by(user_id).pinboard_inserted.where(pinboard_items: {project_id:}).order('updated_at DESC')
     end
 
     h
   end
 
-  protected
+    protected
 
-  # @return [Integer, Nil]
+    # @return [Integer, Nil]
   def extract_tw_attributes
     # NOTE: assumes content type is an image.
     tempfile = image_file.queued_for_write[:original]
@@ -383,8 +386,8 @@ class Image < ApplicationRecord
     end
   end
 
-  # Check md5 fingerprint against existing fingerprints
-  # @return [Object]
+    # Check md5 fingerprint against existing fingerprints
+    # @return [Object]
   def sv_duplicate_image?
     if has_duplicate?
       soft_validations.add(
@@ -393,13 +396,15 @@ class Image < ApplicationRecord
     end
   end
 
-  private
+    private
 
-  # Converts image to blob and releases memory of img (image cannot be used afterwards)
-  # @param [Magick::Image] img
-  # @return [String] a JPG representation of the image
-  # !! Always converts to .jpg, this may need abstraction later
+    # Converts image to blob and releases memory of img (image cannot be used afterwards)
+    # @param [Magick::Image] img
+    # @return [String] a JPG representation of the image
+    #   !! Always converts to .jpg, this may need abstraction later
+    #   Returns an empty string if no image
   def self.to_blob!(img)
+    return '' if img.nil?
     img.format = 'jpg'
     blob = img.to_blob
     img.destroy!
@@ -411,10 +416,10 @@ class Image < ApplicationRecord
 
     dimensions = Paperclip::Geometry.from_file(original)
 
-    errors.add(:image_file, "width must be at least 16 pixels") if dimensions.width < 16
-    errors.add(:image_file, "height must be at least 16 pixels") if dimensions.height < 16
+    errors.add(:image_file, 'width must be at least 16 pixels') if dimensions.width < 16
+    errors.add(:image_file, 'height must be at least 16 pixels') if dimensions.height < 16
   rescue
-    errors.add(:image_file, "unable to extract image dimensions")
+    errors.add(:image_file, 'unable to extract image dimensions')
   end
 
 end
